Reduce image size from 16.5GB to 2.5GB.
To reduce the image size, I adopted the Multi-stage builds approach, splitting the process into two stages: the build stage and the release stage. In the build stage, I followed the same procedure as before, installing all necessary packages and performing the compilation. The release stage, the critical component, involves solely copying the 'install/' folder from the build stage. This streamlined approach is possible because ROS2 only requires the 'install/' folder for execution.

It's worth noting that the build process is time-consuming. I highly recommend avoiding re-building the image and refraining from modifying the Dockerfile unless you encounter a bug. I've already uploaded the image to Dockerhub, so downloading it would be the most efficient option.

Reference
- https://docs.docker.com/build/building/multi-stage/