Add ros2 environment variables.
In the documentation of ROS2-humble, the logging output is colorized by default when targeting a terminal. However, it will not be colorized when I use `osrf/ros:humble-desktop-full-jammy` as a base image. It is fine in the current base image, which is without the `jammy` at the end. I'm not sure why it happened.

I added the environment variable in docker-compose to force enable the functionality in case the osrf modifies the base image in the future.

Reference: https://docs.ros.org/en/humble/Tutorials/Demos/Logging-and-logger-configuration.html#console-output-colorizing