Mount to the top folder
Due to the inconvenience of using a Git container, I decided to modify the mount folder in the Docker Compose configuration.

However, with this setup, there's a potential risk of inadvertently modifying the wrong workspace,
as both the .git folder and other workspace contents are mounted into the container,
which contains many similar elements such as Dockerfiles and docker-compose files.

Similar topic: https://github.com/j3soon/ros2-agv-essentials/pull/13