Fix camera failed to open in waffle model
I've forgot to add robot description in the launch file previous,
here we simply add the robot descriptor.

But in this part we've to use 'str' type as input for opening the urdf file,
so that there existing multiple usage for join a path and get shared package.

I think there may be other way to do the same thing, but it may not be the focus for this branch.