Fix ekf_node not fusing IMU data issue due to topic name mismatch
I choose to rename `imu/data_raw` to `imu/data`. This is because data published by the gazebo imu sensor plugin includes orientation data, which means it is already post-processed.

Although `imu/data_raw` is not required in simulation, it'll be used in real-world deployment. The `imu_filter_madgwick` package is used to convert raw IMU data (without orientation) into a post-processed IMU data (with orientation).

Refs:
* https://github.com/husky/husky/issues/113
* http://wiki.ros.org/imu_filter_madgwick

I believe this topic name mismatch issue is due to the following commit:

Ref: https://github.com/husky/husky/commit/e379ee8a558b20765d36c149a1361bc24c9836d8