Remove the fusion of pose into ekf_node.
To address the issue of pose data from slam_toolbox occasionally jumping over time, it's advisable not to directly fuse it into ekf_node.

In order to prevent the covariance from growing excessively, we can add configuration for `dynamic_process_noise_covariance` in the `robot_localization` package. This will enable the covariance to be dynamically scaled.

Reference:
- https://docs.ros.org/en/melodic/api/robot_localization/html/state_estimation_nodes.html#dynamic-process-noise-covariance

Co-authored-by: Johnson Sun <j3.soon777@gmail.com>