fix(kobuki_ws): Add `COLCON_IGNORE` file inside the docker folder.
Since the `colcon build` command searches the root workspace recursively, it compiles not only the packages in the `src` folder but also the entire workspace. When building other packages in `kobuki_ws`, `colcon build` will also build the driver if it remains inside the root workspace. To prevent this, we should add a `COLCON_IGNORE` file to the `kobuki_driver_ws` directory, and remove it after coping the workspace into the container. At the end, `colcon` will ignore the `kobuki_driver_ws` directory inside the docker folder when building `kobuki_ws`.