fix(template_ws): Allow `rosdep` to continue installing packages despite errors.
Sometimes we might use packages that are no longer maintained. Although these packages can compile normally, `rosdep` might report errors because the `package.xml` is not updated. Using this flag can help avoid this issue. If a package is missing, it will still be displayed in the terminal, so you don’t need to worry about missing error messages.