feat(aloha_ws): Support Vulkan apps such as `vkcube`
These commands are also tested on a fresh machine with NVIDIA GPU on the
`ubuntu:22.04` base image.

When executing `vkcube`, the following error may occur if the Vulkan
config files are missing:

    vkEnumerateInstanceExtensionProperties failed to find the VK_KHR_surface extension.

    Do you have a compatible Vulkan installable client driver (ICD) installed?
    Please look at the Getting Started guide for additional information.

If the required capabilities or libraries are missing, a black window
may popup with the following unwanted message:

    Selected GPU 0: llvmpipe (LLVM 15.0.7, 256 bits), type: 4

It is worth noting that the behavior may differ if the host machine has
Vulkan SDK installed. For example, using `--device /dev/dri` on a host
machine with Vulkan installed may not require certain config files or
libraries. Therefore, it is important to always test on a fresh machine.