fix(template_ws): Fix volume mount permission issue by fixing incorrect bash scripts
Previous commit 96d8787 doesn't fully fix the issue.

This is because in bash scripts with `set -e` flag enabled, the command format `a && b && c` will not halt the script if command `b` fails.

This can be fixed by either using `a; b; c` or `a && b && c || exit 1`, we use the latter one for now to prevent future mistakes.

Ref: https://stackoverflow.com/q/25794905