feat(go2_ws): Update Go2 urdf model.
The `xacro/robot.xacro` and `urdf/go2_description.urdf` are not synchronized, such as the position of IMU is different, there is no head and radar in `robot.xacro`. Based on experience, the urdf is closer to real robot, so `robot.xacro` is updated to be the same as `go2_description.urdf`.

The gazebo plugin is not updated yet.

Note: For easy reading, I rearranged the link and joint order in the order of link first and then joint. In addition, the original `robot.xacro` has been formatted; I formatted it again after modification. Since the difference is not very large, I merged it into one commit directly.